change read_data to work with T

+++ src/cnn/cnn_main.cpp
@@ -7,7 +7,7 @@
 
 using half_float::half;
 int main() {
-  CNN<half>::run();
+  CNN<float>::run();
   // CNN<half>::inference();
   return 0;
 }
